LayoutTests/http/tests/security/xssAuditor/iframe-onload-in-svg-tag-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

 Test that dangerous attributes are still filtered in netsted script contexts.
